requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
alembic==1.18.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

pydantic==2.6.4
python-dotenv==1.0.1

uvicorn==0.41.0